Prevent parallel execution
Note: The check in RunTestCommandHandler is not Thread safe. But the the only
impact is a missing dialog. The lock implementation in TestResourceExecutor
(and its implementations) ensures that multiple executions are executed one at
time. So even if the call to isLocked returns false and a second executions
start before the current execution can get hold of the lock both executions
 are executed on after the other.